refactor(mknod): remove redundant cast in MODE_RW_UGO constant
The bitwise OR operation on the file permission flags already yields a u32 value, making the explicit cast unnecessary and improving code clarity.

+++ src/uu/mknod/src/mknod.rs
@@ -15,7 +15,7 @@ use uucore::format_usage;
 use uucore::fs::makedev;
 use uucore::translate;
 
-const MODE_RW_UGO: u32 = (S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH) as u32;
+const MODE_RW_UGO: u32 = S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH;
 
 mod options {
     pub const MODE: &str = "mode";
